services/template_fetcher/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
xmlschema==3.3.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
torch==2.3.1


